move time utils
